+++ lib/ooxml_parser/common_parser/common_data/alternate_content/drawing/drawing_properties/ooxml_coordinates.rb
@@ -26,7 +26,8 @@ module OoxmlParser
     # @param delimiter [Float] delimiter to devise values
     # @return [OOXMLCoordinates] result of parsing
     def self.parse(position_node, x_attr: 'x', y_attr: 'y', delimiter: OoxmlParser.configuration.units_delimiter)
-      OOXMLCoordinates.new((position_node.attribute(x_attr).value.to_f / delimiter).round(3), (position_node.attribute(y_attr).value.to_f / delimiter).round(3))
+      OOXMLCoordinates.new((position_node.attribute(x_attr).value.to_f / delimiter).round(OoxmlParser.configuration.accuracy),
+                           (position_node.attribute(y_attr).value.to_f / delimiter).round(OoxmlParser.configuration.accuracy))
     end
   end
 end